refactor: move URL checks outside DeploymentAgentService to allow for testing
